* BuildConnectPacket()에서 SessionKey를 송신하지 않도록 수정 * 조용히 끊긴 상태를 확인하기 위한 태스크 추가   * CancellationToken 추가

+++ MultiSocketRUDPBotTester/MultiSocketRUDPBotTester/ClientCore/BotTesterCore.cs
@@ -47,11 +47,11 @@ namespace MultiSocketRUDPBotTester.ClientCore
             }
         }
 
-        public UInt16 GetActiveBotCount()
+        public int GetActiveBotCount()
         {
             lock (sessionDictionaryLock)
             {
-                return (UInt16)sessionDictionary.Count;
+                return sessionDictionary.Count;
             }
         }
